requirements.txt
django==4.2
pillow==11.3.0
sqlparse==0.5.3
dotenv==0.9.9
psycopg==3.2.10
redis==7.0.1
pyjwt==2.10.1